validate delete

+++ app/Http/Controllers/AgentController.php
@@ -107,6 +107,10 @@ class AgentController extends Controller
             return $this->renderError($request, __("authorize.not_superadmin"), 401);
         }
 
+        if (!empty($agent->transactions)) {
+            return $this->renderError($request, __("rules.cannot_delete_agent_has_transaction"), 422);
+        }
+
         $agent->delete();
         return $this->renderView($request, '', [], ['route' => 'agents.index', 'data' => []], 204);
     }
